RAG/Python/basic-rag_openai-ingest-query/requirements.txt
llama-index>=0.10.0
llama-index-vector-stores-chroma>=0.1.0
llama-index-embeddings-openai>=0.1.0
llama-index-llms-openai>=0.1.0
chromadb>=0.4.0
openai>=1.0.0
python-dotenv>=1.0.0
streamlit>=1.32.0
tiktoken>=0.6.0
